+++ modules/methods/src/com/fs/starfarer/settings/StarfarerSettings.java
@@ -1,0 +1,13 @@
+package com.fs.starfarer.settings;
+
+public class StarfarerSettings {
+    // $FF: renamed from: ö00000 (java.lang.String) float
+    public static float method_17047(String var0) {
+        return 0;
+    }
+
+    // ADDED METHOD
+    public static float getFPS() {
+        return method_17047("fps");
+    }
+}